+++ ui/src/NavigationBar/NavigationBar.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import { ThemeContext } from "../context/ThemeContext";
-import { SmoothSvgArch } from "./SmoothSvgArch";
+import { SvgArch } from "./SvgArch";
 import { SidebarOverlay } from "../Sidebar/Sidebar";
 
 interface NavButtons {
@@ -23,10 +23,25 @@ export function NavigationBar() {
     throw new Error("DarkMode must be used within a ThemeProvider");
   }
   const { isDarkMode } = themeContext;
+    const [scrollY, setScrollY] = useState(0);
+
+  useEffect(() => {
+    const handleScroll = () => setScrollY(window.scrollY);
+
+    window.addEventListener("scroll", handleScroll);
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, []);
   return (
     <>
-      <div className="navBar__long">
-        <Link to="/home"><img src="images/Pillow_Logo.png" alt="pill_logo" className="navBar__long__logo"/></Link>
+      <div className="navBar__long"
+        style={{
+          flexDirection: `${scrollY <= 100 ? "column" : "row"}`,
+          position: `${scrollY <= 100 ? "absolute" : "fixed"}`,
+          }}>
+        <Link to="/home"><img src="images/Pillow_Logo.png" alt="pill_logo" className="navBar__long__logo"
+          style={{
+            width: `${scrollY <= 100 ? "175px" : "150px"}`,
+          }} /></Link>
         <div className="navBar__long__content">
           {navButtons.map((button) => (
             <Link
@@ -35,6 +50,10 @@ export function NavigationBar() {
                 isDarkMode ? "navBar__long__button__night" : "navBar__long__button"
               }
               key={button.label}
+            style={{
+              paddingLeft: `${scrollY <= 100 ? "69px" : "69px"}`,
+              paddingRight: `${scrollY <= 100 ? "69px" : "69px"}`
+            }}
             >
               <span className="navBar__long__button__label">{button.label}</span>
             </Link>
@@ -42,11 +61,11 @@ export function NavigationBar() {
         </div>
       </div>
       <div className="navBar__short">
-        <Link to="/home"><img src="images/Pillow_Logo.png" alt="pill_logo" className="navBar__short__logo"/></Link>
+        <Link to="/home"><img src="images/Pillow_Logo.png" alt="pill_logo" className="navBar__short__logo" /></Link>
+
         <SidebarOverlay />
-      
       </div>
-      <SmoothSvgArch colorMode={isDarkMode} />
+      <SvgArch colorMode={isDarkMode} />
     </>
   );
 }
